+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.1.1 (unreleased)
 
+- Improved performance of loading models
 - Fixed error with `aggregation_strategy` option
 
 ## 0.1.0 (2024-08-19)